Implement credentials UI for campaigns user tokens (#15506)
This adds an implementation of the GraphQL API resolvers and the UI part of the campaigns user tokens project.

+++ cmd/frontend/graphqlbackend/campaigns.go
@@ -87,7 +87,7 @@ type DeleteCampaignsCredentialArgs struct {
 type ListCampaignsCodeHostsArgs struct {
 	First  int32
 	After  *string
-	UserID int64
+	UserID int32
 }
 
 type CampaignsResolver interface {
